requirements.txt
alembic==0.9.5
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
json-rpc==1.10.8
kappa==0.6.0
lambda-packages==0.19.0
Mako==1.0.7
MarkupSafe==1.0
pbkdf2==1.3
placebo==0.8.1
pycparser==2.18
pycryptodome==3.4.7
pyethash==0.1.27
pylru==1.0.9
pysha3==1.0.2
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
s3transfer==0.1.12
scrypt==0.8.0
secp256k1==0.13.2
semantic-version==2.6.0
six==1.11.0
toolz==0.8.2
tqdm==4.19.1
troposphere==2.1.2
Unidecode==0.4.21
urllib3==1.22
web3==3.16.3
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
